Type Definition atsamd11c::dmac::chctrlb::TRIGSRC_R

type TRIGSRC_R = R<u8, TRIGSRC_A>;

Reader of field TRIGSRC

Implementations

impl TRIGSRC_R

pub fn variant(&self) -> Variant<u8, TRIGSRC_A>

Get enumerated values variant

pub fn is_disable(&self) -> bool

Checks if the value of the field is DISABLE

pub fn is_sercom0_rx(&self) -> bool

Checks if the value of the field is SERCOM0_RX

pub fn is_sercom0_tx(&self) -> bool

Checks if the value of the field is SERCOM0_TX

pub fn is_sercom1_rx(&self) -> bool

Checks if the value of the field is SERCOM1_RX

pub fn is_sercom1_tx(&self) -> bool

Checks if the value of the field is SERCOM1_TX

pub fn is_sercom2_rx(&self) -> bool

Checks if the value of the field is SERCOM2_RX

pub fn is_sercom2_tx(&self) -> bool

Checks if the value of the field is SERCOM2_TX

pub fn is_tcc0_ovf(&self) -> bool

Checks if the value of the field is TCC0_OVF

pub fn is_tcc0_mc0(&self) -> bool

Checks if the value of the field is TCC0_MC0

pub fn is_tcc0_mc1(&self) -> bool

Checks if the value of the field is TCC0_MC1

pub fn is_tcc0_mc2(&self) -> bool

Checks if the value of the field is TCC0_MC2

pub fn is_tcc0_mc3(&self) -> bool

Checks if the value of the field is TCC0_MC3

pub fn is_tc1_ovf(&self) -> bool

Checks if the value of the field is TC1_OVF

pub fn is_tc1_mc0(&self) -> bool

Checks if the value of the field is TC1_MC0

pub fn is_tc1_mc1(&self) -> bool

Checks if the value of the field is TC1_MC1

pub fn is_tc2_ovf(&self) -> bool

Checks if the value of the field is TC2_OVF

pub fn is_tc2_mc0(&self) -> bool

Checks if the value of the field is TC2_MC0

pub fn is_tc2_mc1(&self) -> bool

Checks if the value of the field is TC2_MC1

pub fn is_adc_resrdy(&self) -> bool

Checks if the value of the field is ADC_RESRDY

pub fn is_dac_empty(&self) -> bool

Checks if the value of the field is DAC_EMPTY