Type Definition atsamd11c::eic::config::SENSE4_R

type SENSE4_R = R<u8, SENSE4_A>;

Reader of field SENSE4

Implementations

impl SENSE4_R

pub fn variant(&self) -> Variant<u8, SENSE4_A>

Get enumerated values variant

pub fn is_none(&self) -> bool

Checks if the value of the field is NONE

pub fn is_rise(&self) -> bool

Checks if the value of the field is RISE

pub fn is_fall(&self) -> bool

Checks if the value of the field is FALL

pub fn is_both(&self) -> bool

Checks if the value of the field is BOTH

pub fn is_high(&self) -> bool

Checks if the value of the field is HIGH

pub fn is_low(&self) -> bool

Checks if the value of the field is LOW