Type Definition atsamd11c::gclk::clkctrl::ID_R

type ID_R = R<u8, ID_A>;

Reader of field ID

Implementations

impl ID_R

pub fn variant(&self) -> Variant<u8, ID_A>

Get enumerated values variant

pub fn is_dfll48(&self) -> bool

Checks if the value of the field is DFLL48

pub fn is_fdpll(&self) -> bool

Checks if the value of the field is FDPLL

pub fn is_fdpll32k(&self) -> bool

Checks if the value of the field is FDPLL32K

pub fn is_wdt(&self) -> bool

Checks if the value of the field is WDT

pub fn is_rtc(&self) -> bool

Checks if the value of the field is RTC

pub fn is_eic(&self) -> bool

Checks if the value of the field is EIC

pub fn is_usb(&self) -> bool

Checks if the value of the field is USB

pub fn is_evsys_0(&self) -> bool

Checks if the value of the field is EVSYS_0

pub fn is_evsys_1(&self) -> bool

Checks if the value of the field is EVSYS_1

pub fn is_evsys_2(&self) -> bool

Checks if the value of the field is EVSYS_2

pub fn is_evsys_3(&self) -> bool

Checks if the value of the field is EVSYS_3

pub fn is_evsys_4(&self) -> bool

Checks if the value of the field is EVSYS_4

pub fn is_evsys_5(&self) -> bool

Checks if the value of the field is EVSYS_5

pub fn is_sercomx_slow(&self) -> bool

Checks if the value of the field is SERCOMX_SLOW

pub fn is_sercom0_core(&self) -> bool

Checks if the value of the field is SERCOM0_CORE

pub fn is_sercom1_core(&self) -> bool

Checks if the value of the field is SERCOM1_CORE

pub fn is_sercom2_core(&self) -> bool

Checks if the value of the field is SERCOM2_CORE

pub fn is_tcc0(&self) -> bool

Checks if the value of the field is TCC0

pub fn is_tc1_tc2(&self) -> bool

Checks if the value of the field is TC1_TC2

pub fn is_adc(&self) -> bool

Checks if the value of the field is ADC

pub fn is_ac_dig(&self) -> bool

Checks if the value of the field is AC_DIG

pub fn is_ac_ana(&self) -> bool

Checks if the value of the field is AC_ANA

pub fn is_dac(&self) -> bool

Checks if the value of the field is DAC