Type Definition atsamd11c::nvmctrl::ctrla::CMD_R

type CMD_R = R<u8, CMD_A>;

Reader of field CMD

Implementations

impl CMD_R

pub fn variant(&self) -> Variant<u8, CMD_A>

Get enumerated values variant

pub fn is_er(&self) -> bool

Checks if the value of the field is ER

pub fn is_wp(&self) -> bool

Checks if the value of the field is WP

pub fn is_ear(&self) -> bool

Checks if the value of the field is EAR

pub fn is_wap(&self) -> bool

Checks if the value of the field is WAP

pub fn is_sf(&self) -> bool

Checks if the value of the field is SF

pub fn is_wl(&self) -> bool

Checks if the value of the field is WL

pub fn is_lr(&self) -> bool

Checks if the value of the field is LR

pub fn is_ur(&self) -> bool

Checks if the value of the field is UR

pub fn is_sprm(&self) -> bool

Checks if the value of the field is SPRM

pub fn is_cprm(&self) -> bool

Checks if the value of the field is CPRM

pub fn is_pbc(&self) -> bool

Checks if the value of the field is PBC

pub fn is_ssb(&self) -> bool

Checks if the value of the field is SSB

pub fn is_invall(&self) -> bool

Checks if the value of the field is INVALL