atsamd11c::nvmctrl::ctrla

Type Alias CmdW

pub type CmdW<'a, REG> = FieldWriter<'a, REG, 7, Cmdselect>;

Field CMD writer - Command

Aliased Type

struct CmdW<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> CmdW<'a, REG>
where REG: Writable + RegisterSpec, REG::Ux: From<u8>,

pub fn er(self) -> &'a mut W<REG>

Erase Row - Erases the row addressed by the ADDR register.

pub fn wp(self) -> &'a mut W<REG>

Write Page - Writes the contents of the page buffer to the page addressed by the ADDR register.

pub fn ear(self) -> &'a mut W<REG>

Erase Auxiliary Row - Erases the auxiliary row addressed by the ADDR register. This command can be given only when the security bit is not set and only to the user configuration row.

pub fn wap(self) -> &'a mut W<REG>

Write Auxiliary Page - Writes the contents of the page buffer to the page addressed by the ADDR register. This command can be given only when the security bit is not set and only to the user configuration row.

pub fn sf(self) -> &'a mut W<REG>

Security Flow Command

pub fn wl(self) -> &'a mut W<REG>

Write lockbits

pub fn lr(self) -> &'a mut W<REG>

Lock Region - Locks the region containing the address location in the ADDR register.

pub fn ur(self) -> &'a mut W<REG>

Unlock Region - Unlocks the region containing the address location in the ADDR register.

pub fn sprm(self) -> &'a mut W<REG>

Sets the power reduction mode.

pub fn cprm(self) -> &'a mut W<REG>

Clears the power reduction mode.

pub fn pbc(self) -> &'a mut W<REG>

Page Buffer Clear - Clears the page buffer.

pub fn ssb(self) -> &'a mut W<REG>

Set Security Bit - Sets the security bit by writing 0x00 to the first byte in the lockbit row.

pub fn invall(self) -> &'a mut W<REG>

Invalidates all cache lines.