Type Definition atsamd11c::nvmctrl::param::PSZ_R

type PSZ_R = R<u8, PSZ_A>;

Reader of field PSZ

Implementations

impl PSZ_R

pub fn variant(&self) -> PSZ_A

Get enumerated values variant

pub fn is_8(&self) -> bool

Checks if the value of the field is _8

pub fn is_16(&self) -> bool

Checks if the value of the field is _16

pub fn is_32(&self) -> bool

Checks if the value of the field is _32

pub fn is_64(&self) -> bool

Checks if the value of the field is _64

pub fn is_128(&self) -> bool

Checks if the value of the field is _128

pub fn is_256(&self) -> bool

Checks if the value of the field is _256

pub fn is_512(&self) -> bool

Checks if the value of the field is _512

pub fn is_1024(&self) -> bool

Checks if the value of the field is _1024