atsamd11c::tcc0::evctrl

Type Alias CntselW

pub type CntselW<'a, REG> = FieldWriter<'a, REG, 2, Cntselselect, Safe>;

Field CNTSEL writer - Timer/counter Output Event Mode

Aliased Type

struct CntselW<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> CntselW<'a, REG>
where REG: Writable + RegisterSpec, REG::Ux: From<u8>,

pub fn start(self) -> &'a mut W<REG>

An interrupt/event is generated when a new counter cycle starts

pub fn end(self) -> &'a mut W<REG>

An interrupt/event is generated when a counter cycle ends

pub fn between(self) -> &'a mut W<REG>

An interrupt/event is generated when a counter cycle ends, except for the first and last cycles

pub fn boundary(self) -> &'a mut W<REG>

An interrupt/event is generated when a new counter cycle starts or a counter cycle ends